// projects/series-movies/src/app/models/enum/genre.enum.ts
export enum GenreType {
  ADVENTURE = "adventure",
  ACTION = "action",
  ANIMATION = "animation",
  COMEDY = "comedy",
  CRIME = "crime",
  DOCUMENTARY = "documentary",
  DRAMA = "drama",
  FANTASY = "fantasy",
  HORROR = "horror",
  COOKING_SHOW = "cooking-show",
  LGBTQ = "lgbtq",
  ANIME = "anime",
  MYSTERY = "mystery",
  PODCAST = "podcast",
  QUIZ = "quiz",
  ROMANCE = "romance",
  SCIENCE_FICTION = "science-fiction",
  SPORT = "sport",
  THRILLER = "thriller",
  TV_SHOW = "tv-show",
  TALKSHOW = "talkshow",
  GAMING = "gaming",
  TEENIE = "teenie",
  POLITICS = "politics",
  NEWS = "news",
  SPY = "spy",
}